Refactor routes and controllers to pass the req, res, next objects (#20)

+++ controllers/tournament.controller.js
@@ -8,18 +8,28 @@ const db = require('../helpers/db');
 // const tokenService = require('services/token.service');
 const TournamentModel = db.Tournament;
 
-async function create(tournamentParam) {
+async function create(req, res, next) {
+	const tournamentParam = req.body;
+
 	if (!tournamentParam || !tournamentParam.name || !tournamentParam.clubId) {
-		throw new Error('Invalid parameters');
+		return next(Error('Invalid parameters'));
 	}
 
 	const savedTournament = await TournamentModel.create(tournamentParam);
 
-	return { tournament: savedTournament };
+	return res.status(200).json({ tournament: savedTournament });
 }
 
-async function deleteTournament(tournamentId) {
+async function deleteTournament(req, res, next) {
+	const tournamentId = req.params.id;
+
+	if (!tournamentId) {
+		return next(Error('Invalid parameters'));
+	}
+
 	await TournamentModel.deleteById(tournamentId);
+
+	return res.status(200).json({ message: 'success' });
 }
 
 module.exports = {
